+++ main.go
@@ -1,11 +1,12 @@
 package main
 
-import sitemap_builder "goPhercise/sitemap-builder"
+import aecho "goPhercise/goBook/1-A-Echo"
 
 func main() {
 	//quiz_game.RollQuizGame()
 	//url_shortener.RollURLShortener()
 	//cya_game.RollCYAGame()
 	//HTML Link parser is run in its own directories
-	sitemap_builder.RollSitemapBuilder()
+	aecho.RollEcho()
+	//sitemap_builder.RollSitemapBuilder()
 }